\name{LungCancerACvsSCCGEO}
\alias{LungCancerACvsSCCGEO}
\alias{anoLC}
\alias{gsLC}



\title{Annotation of a small set of training and test set samples (30 total) used by team 221 in the IMPROVER DSC
 for the lung cancer sub-challenge.}

\description{
  The \code{LungCancerACvsSCCGEO} dataset consists: i) a data frame \code{anoLC} giving
  the file names of the affy cel files used in the training phase and their corresponding 
  phenotype (AC or SCC) and ii) \code{gsLC} the gold standard, i.e. the class membership
  of each test sample apearing in \code{anoLC}. 
 }

\usage{data(LC)}

\source{ GEO for the training data, while the test data comes from the citation below: 
Adi L. Tarca, Mario Lauria, Michael Unger, Erhan Bilal, Stephanie Boue, Kushal Kumar Dey,
 Julia Hoeng, Heinz Koeppl, Florian Martin, Pablo Meyer, Preetam Nandy, Raquel Norel,
  Manuel Peitsch, Jeremy J Rice, Roberto Romero, Gustavo Stolovitzky, Marja Talikka,
   Yang Xiang, Christoph Zechner, and IMPROVER DSC Collaborators, 
   Strengths and limitations of microarray-based phenotype prediction: Lessons learned 
   from the IMPROVER Diagnostic Signature Challenge. Bioinformatics, submitted 2013. \cr
}

\keyword{datasets}


